+++ src/main/java/racingGame/domain/utils/RandomUtils.java
@@ -1,4 +1,4 @@
-package racingcar.domain.utils;
+package racingGame.domain.utils;
 
 import java.util.Random;
