+++ src/main/java/racingGame/domain/validator/InputValidator.java
@@ -1,4 +1,4 @@
-package racingcar.domain.validator;
+package racingGame.domain.validator;
 
 public class InputValidator {
