+++ src/test/java/racingGame/domain/utils/NameHandlerTest.java
@@ -1,9 +1,9 @@
-package racingcar.domain.utils;
+package racingGame.domain.utils;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.domain.Name;
+import racingGame.domain.racingCar.Name;
 
 import java.util.List;
